requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==23.2.0
Automat==22.10.0
buildtools==1.0.6
cachetools==5.2.0
causal-conv1d==1.2.0.post2
ccimport
charset-normalizer==3.3.2
constantly==23.10.4
cumm-cu120
docopt==0.6.2
dpcpp-cpp-rt==2024.0.3
easydict==1.13
einops==0.7.0
filelock==3.13.1
fire==0.6.0
flatbuffers==1.12
fsspec==2024.3.1
furl==2.1.3
fvcore
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.34.1
h5py
huggingface-hub==0.21.4
hyperlink==21.0.0
idna==3.6
imageio
importlib_metadata==7.0.2
incremental==22.10.0
intel-cmplr-lib-rt==2024.0.3
intel-cmplr-lic-rt==2024.0.3
intel-opencl-rt==2024.0.3
intel-openmp==2024.0.3
iopath==0.1.10
Jinja2==3.1.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing
lark==1.1.9
lazy_loader==0.3
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
mkl==2024.0.0
mkl-fft
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numba==0.56.4
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==24.0
pccm
pillow==10.2.0
portalocker==2.8.2
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
redo==2.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.19.3
scipy==1.9.1
SharedArray==3.0.0
simplejson==3.19.2
six==1.15.0
spconv-cu120==2.3.6
SQLAlchemy
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tifffile==2023.7.10
timm==0.9.16
tokenizers==0.15.2
torch-scatter==2.1.2
tqdm==4.66.2
transformers==4.38.2
triton==2.1.0
Twisted
typing_extensions==4.10.0
urllib3==2.2.1
waymo-open-dataset-tf-2-5-0==1.4.1
Werkzeug==2.2.2
wrapt==1.12.1
yacs==0.1.8
zipp==3.18.1
zope.interface==6.2